+++ src/osc/wave_table_osc.rs
@@ -6,23 +6,23 @@ use crate::util::Component;
 use crate::util::WAVE_TABLE_SAMPLES_PER_CYCLE;
 
 pub struct WaveTableOsc {
-    pub counter: u16,
-    pub wt: [i8; WAVE_TABLE_SAMPLES_PER_CYCLE as usize],
-    pub freq_ipc: u16,
-    pub freq: i8,
-    pub modulation_idx: i8,
-    pub modulation: i8,
-    pub phase_offset: i8,
-    pub out_cv: i8,
-    pub dummy: i8,
-    ipc_64_map: [u16; 256],
+    pub counter: u32,
+    pub wt: [i16; WAVE_TABLE_SAMPLES_PER_CYCLE as usize],
+    pub freq_ipc: u32,
+    pub freq: i16,
+    pub modulation_idx: i16,
+    pub modulation: i16,
+    pub phase_offset: i16,
+    pub out_cv: i16,
+    pub dummy: i16,
+    ipc_64_map: [u32; 256],
 }
 
 impl WaveTableOsc {
     pub fn new(
-        wt: [i8; WAVE_TABLE_SAMPLES_PER_CYCLE as usize],
-        ipc_64_map: [u16; 256],
-        init_freq: i8,
+        wt: [i16; WAVE_TABLE_SAMPLES_PER_CYCLE as usize],
+        ipc_64_map: [u32; 256],
+        init_freq: i16,
     ) -> WaveTableOsc {
         WaveTableOsc {
             counter: 0,
@@ -38,7 +38,7 @@ impl WaveTableOsc {
         }
     }
 
-    pub fn sin(ipc_64_map: [u16; 256], init_freq: i8) -> WaveTableOsc {
+    pub fn sin(ipc_64_map: [u32; 256], init_freq: i16) -> WaveTableOsc {
         let mut wto = WaveTableOsc::new(
             [0; WAVE_TABLE_SAMPLES_PER_CYCLE as usize],
             ipc_64_map,
@@ -48,26 +48,26 @@ impl WaveTableOsc {
             wto.wt[i as usize] = ((((i as f64) / (WAVE_TABLE_SAMPLES_PER_CYCLE as f64)
                 * std::f64::consts::TAU)
                 .sin())
-                * (i8::max_value() as f64)) as i8;
+                * (i16::max_value() as f64)) as i16;
         }
 
         wto
     }
-    pub fn saw(ipc_64_map: [u16; 256], init_freq: i8) -> WaveTableOsc {
+    pub fn saw(ipc_64_map: [u32; 256], init_freq: i16) -> WaveTableOsc {
         let mut wto = WaveTableOsc::new(
             [0; WAVE_TABLE_SAMPLES_PER_CYCLE as usize],
             ipc_64_map,
             init_freq,
         );
         for i in 0..WAVE_TABLE_SAMPLES_PER_CYCLE as usize {
-            wto.wt[i] = ((i8::max_value() as f64)
-                - ((u8::max_value() as f64) * ((i as f64) / (WAVE_TABLE_SAMPLES_PER_CYCLE as f64))))
-                as i8;
+            wto.wt[i] = ((i16::max_value() as f64)
+                - ((u16::max_value() as f64) * ((i as f64) / (WAVE_TABLE_SAMPLES_PER_CYCLE as f64))))
+                as i16;
         }
 
         wto
     }
-    pub fn triangle(ipc_64_map: [u16; 256], init_freq: i8) -> WaveTableOsc {
+    pub fn triangle(ipc_64_map: [u32; 256], init_freq: i16) -> WaveTableOsc {
         let mut wto = WaveTableOsc::new(
             [0; WAVE_TABLE_SAMPLES_PER_CYCLE as usize],
             ipc_64_map,
@@ -75,33 +75,33 @@ impl WaveTableOsc {
         );
         for i in 0..WAVE_TABLE_SAMPLES_PER_CYCLE as usize {
             if i <= WAVE_TABLE_SAMPLES_PER_CYCLE as usize / 4 {
-                wto.wt[i] = ((i8::max_value() as f64) * (i as f64)
+                wto.wt[i] = ((i16::max_value() as f64) * (i as f64)
                     / ((WAVE_TABLE_SAMPLES_PER_CYCLE as f64) / 4.))
-                    as i8;
+                    as i16;
             } else if i <= (WAVE_TABLE_SAMPLES_PER_CYCLE as usize) / 2 {
-                wto.wt[i] = ((i8::max_value() as f64)
+                wto.wt[i] = ((i16::max_value() as f64)
                     * (1.
                         - ((i as f64) - ((WAVE_TABLE_SAMPLES_PER_CYCLE as f64) / 4.))
                             / ((WAVE_TABLE_SAMPLES_PER_CYCLE as f64) / 4.)))
-                    as i8;
+                    as i16;
             } else if i <= (WAVE_TABLE_SAMPLES_PER_CYCLE as usize) / 4 * 3 {
-                wto.wt[i] = ((i8::min_value() as f64)
+                wto.wt[i] = ((i16::min_value() as f64)
                     * ((i as f64) - ((WAVE_TABLE_SAMPLES_PER_CYCLE as f64) / 2.))
                     / ((WAVE_TABLE_SAMPLES_PER_CYCLE as f64) / 4.))
-                    as i8;
+                    as i16;
             } else {
-                wto.wt[i] = ((i8::min_value() as f64)
+                wto.wt[i] = ((i16::min_value() as f64)
                     * (1.
                         - ((i as f64) - ((WAVE_TABLE_SAMPLES_PER_CYCLE as f64) / 4. * 3.))
                             / ((WAVE_TABLE_SAMPLES_PER_CYCLE as f64) / 4.)))
-                    as i8;
+                    as i16;
             }
         }
 
         wto
     }
 
-    pub fn square(ipc_64_map: [u16; 256], init_freq: i8) -> WaveTableOsc {
+    pub fn square(ipc_64_map: [u32; 256], init_freq: i16) -> WaveTableOsc {
         let mut wto = WaveTableOsc::new(
             [0; WAVE_TABLE_SAMPLES_PER_CYCLE as usize],
             ipc_64_map,
@@ -109,9 +109,9 @@ impl WaveTableOsc {
         );
         for i in 0..(WAVE_TABLE_SAMPLES_PER_CYCLE as usize) {
             if i <= (WAVE_TABLE_SAMPLES_PER_CYCLE as usize) / 2 {
-                wto.wt[i] = i8::max_value();
+                wto.wt[i] = i16::max_value();
             } else {
-                wto.wt[i] = i8::min_value();
+                wto.wt[i] = i16::min_value();
             }
         }
 
@@ -125,8 +125,8 @@ impl<'a> Component<'a> for WaveTableOsc {
         self.freq_ipc = self.ipc_64_map[self.freq as usize];
         // So, in theory we could have a wt with multiple frames in it,
         // so I don't want to hardcode this right now.
-        // let wt_len = (64 * self.wt.len()) as u16;
-        // Setting the len to 1024 allows natural wrapping of a u16.
+        // let wt_len = (64 * self.wt.len()) as u32;
+        // Setting the len to 1024 allows natural wrapping of a u32.
 
         self.counter = self.counter.wrapping_add(self.freq_ipc);
 
@@ -134,13 +134,13 @@ impl<'a> Component<'a> for WaveTableOsc {
         // I am trying to use the modulation_idx as essentially as a signed Q1.7
         //println!("{} {}", self.freq_ipc, self.modulation_idx);
         let m = (self.freq_ipc as i32) * (self.modulation_idx as i32) >> 7;
-        let m = (((self.modulation as i32) * m) >> 7) as i8;
+        let m = (((self.modulation as i32) * m) >> 7) as i16;
 
         // I need to double check that this works the way I'm expecting
         // with the wrapping. Also need to think about how this would
         // be implemented on a microcontroller.
-        self.counter = (self.counter as i32).wrapping_add(m as i32) as u16;
-        // Setting the len to 1024 allows natural wrapping of a u16.
+        self.counter = (self.counter as i32).wrapping_add(m as i32) as u32;
+        // Setting the len to 1024 allows natural wrapping of a u32.
         // self.counter %= wt_len;
 
         // I need to double check that this works the way I'm expecting
@@ -148,7 +148,7 @@ impl<'a> Component<'a> for WaveTableOsc {
         // be implemented on a microcontroller.
         let mut i = (self.counter as i32).wrapping_add(m as i32) as usize;
         i = i.wrapping_add(self.phase_offset as usize);
-        // Setting the len to 1024 allows natural wrapping of a u16.
+        // Setting the len to 1024 allows natural wrapping of a u32.
         i >>= 6;
         i %= self.wt.len();
 
@@ -165,7 +165,7 @@ impl<'a> Component<'a> for WaveTableOsc {
 }
 
 impl Index<&str> for WaveTableOsc {
-    type Output = i8;
+    type Output = i16;
 
     fn index(&self, i: &str) -> &Self::Output {
         match i {
